mainMenu renomé en main_menu
